week-5/day-1/daily_challenge/phone_number/requirements.txt
asgiref==3.6.0
Django==4.2
django-phone-field==1.8.1
psycopg2-binary==2.9.6
sqlparse==0.4.4
